Update returns for error handling
