+++ src/dom-updates.js
@@ -1,14 +1,14 @@
 
 import { use } from "chai";
-import { traveler } from "./data-model";
-import { createPastTripCardElement, getTotalSpentOnTrips } from "./data-model";
+import { fetchData, traveler } from "./data-model";
+import { getTotalSpentOnTrips, getTraveler, getTravelerTrips, getDestinationDetails, getTripTotal } from "./data-model";
 //////domUpdates/////////
 export const displayTrips = () => {
     const cardContainer = document.getElementById('past-trips-container')
     cardContainer.innerHTML = ''
     traveler.trips.forEach(element => {
         if (element.status === 'approved') {
-        const cardElement = createPastTripCardElement(element)
+        const cardElement = createTripCardElement(element)
         cardContainer.appendChild(cardElement)
         }
     });
@@ -20,7 +20,7 @@ export const displayUpcomingTrips = () => {
     cardContainer.innerHTML = ''
     traveler.trips.forEach((element) => {
         if (element.status === 'pending') {
-        const cardElements = createPastTripCardElement(element)
+        const cardElements = createTripCardElement(element)
         cardContainer.appendChild(cardElements)
     }
     })
@@ -29,7 +29,7 @@ export const displayUpcomingTrips = () => {
 export const displayYTDSpend = () => {
     const totalDisplay = document.getElementById('ytd-spend')
     totalDisplay.innerHTML = ''
-    totalDisplay.innerHTML = `Welcome ${traveler.info.name}! <br> -YTD Spend- <br> $${getTotalSpentOnTrips()}`
+    totalDisplay.innerHTML = `-Welcome- <br> ${traveler.info.name}! <br> <br> -YTD Spend- <br> $${getTotalSpentOnTrips()}`
 }
 
 export const displayTripEstimate = (trip) => {
@@ -38,19 +38,89 @@ export const displayTripEstimate = (trip) => {
     form.id = 'hidden-form'
     estimateDisplay.classList.remove('hidden')
     estimateDisplay.innerHTML = ''
-    const cardElement = createPastTripCardElement(trip)
+    const cardElement = createTripCardElement(trip)
     cardElement.classList.add('estimate-display-card')
     estimateDisplay.appendChild(cardElement)
 }
 
+export const createSelectionDestinations = () => {
+    const dropDown = document.getElementById('destination')
+    return fetchData.destinations.forEach((element) => {
+        const option = document.createElement('option')
+        option.value = element.id
+        option.text = element.destination
+        dropDown.appendChild(option)
+    })
+}
+
+export const handleLogIn = (userName, passWord) => {
+    const prefix = 'traveler'
+    if (passWord != 'travel') {
+        return console.log('Incorrect Password')
+    } else if (userName.startsWith(prefix)) {
+        const numValue = userName.substring(prefix.length)
+        const userNum = parseInt(numValue)
+        getTraveler(userNum, fetchData)
+        getTravelerTrips()
+        displayYTDSpend()
+        createSelectionDestinations()
+        traveler.login = true
+    }
+}
 export const handleBackToFormClick = () => {
     const tripEstimateDisplay = document.querySelector('.trip-estimate-display')
     const bookForm = document.getElementById('hidden-form')
-    const clearForm = document.querySelectorAll('.form-remove')
+    // const clearForm = document.querySelectorAll('.form-remove')
+    // console.log(clearForm)
     const estimateCardElement = document.querySelector('.estimate-display-card')
     tripEstimateDisplay.classList.add('hidden')
     tripEstimateDisplay.removeChild(estimateCardElement)
     bookForm.id = "booking-Form"
 }
 
+export const createTripCardElement = (trip) => {
+    const destinationDetails = getDestinationDetails(trip)
+    const country = destinationDetails.destination.split(',')[1]
+    const tripTotal = (getTripTotal(trip)).toLocaleString()
+    const numNights = trip.duration - 1
+    const cardElement = document.createElement('article')
+    if (trip.status === 'approved') {
+        cardElement.classList.add('past-trips')
+        cardElement.innerHTML = `
+        <div class="trip-card-header">
+            <h2 class="trip-card-location">${destinationDetails.destination}</h2>
+            <img src=${destinationDetails.image} alt=${destinationDetails.alt}>
+        </div>
+        <div class="trip-card-body">
+            <p class="trip-details">On ${trip.date}, you spent ${trip.duration} days and ${numNights} nights with ${trip.travelers} guests in beautiful ${country}!</p>
+            <p class="trip-cost">Total: $${tripTotal}</p>
+        </div>
+    `
+    return cardElement
+    } else if (trip.status === 'pending') {
+        cardElement.classList.add('past-trips')
+        cardElement.innerHTML = `
+        <div class="trip-card-header form-remove">
+            <h2 class="trip-card-location">${destinationDetails.destination}</h2>
+            <img class="estimate-pic" src=${destinationDetails.image} alt=${destinationDetails.alt}>
+        </div>
+        <div class="trip-card-body form-remove">
+            <p class="trip-details">Trip Date: ${trip.date}</p>
+            <p class="trip-details">Duration: ${trip.duration}</p>
+            <p class="trip-details">Guests: ${trip.travelers}</p>
+            <p class="trip-details">Status: ${trip.status}</p>
+            <p class="trip-cost">Est. Trip Total: $${tripTotal}</p>
+        </div>
+        `
+        return cardElement
+    }
+
+}
+
+// export const clearForm = () => {
+//     dateInput.value = '' 
+//     numNightsInput.value = ''
+//     numGuestInput.value = ''
+//     destinationSelection.value = ''
+// }
